# backend/openapi/nix-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Nix API",
    "description": "Contract for the Nix collaborative document workspace. The generated frontend client and its MSW mocks are built from this document.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/items/{itemId}/schema": {
      "get": {
        "tags": [
          "Structure"
        ],
        "summary": "The property schema in force at an item",
        "description": "Returns the merged result of every ancestor's declaration, nearest winning, alongside the subset this item declares itself. Both are needed: an editor shown only the merged result would save inherited properties back onto the item and silently turn inheritance into a copy.",
        "operationId": "GetEffectiveSchema",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Structure"
        ],
        "summary": "Declare the property schema for an item's subtree",
        "description": "Replaces what this item declares. Ancestors are unaffected, and so are the property values already stored beneath it: a property removed from a schema stops being validated and stops being displayed, and returns intact if the schema does. Fails with 'schema.invalid' when the document cannot be stored.",
        "operationId": "SetItemSchema",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSchemaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{itemId}/properties": {
      "patch": {
        "tags": [
          "Structure"
        ],
        "summary": "Write property values onto an item",
        "description": "Merges the supplied properties into the item's bag: a member set to null clears that property, and anything not mentioned is left alone. This is what a board drag and a calendar drag both perform - the change is to the item, so it is visible in every view. Fails with 'properties.invalid' when a value does not fit the schema in force.",
        "operationId": "SetItemProperties",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPropertiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{itemId}/views": {
      "get": {
        "tags": [
          "Structure"
        ],
        "summary": "The views a container offers",
        "description": "Returns the views in switcher order, plus the identifiers of any whose configured property no longer exists or no longer fits. A board grouping by a deleted property would otherwise render as an empty board, which is indistinguishable from an item with nothing in it. A list and a gallery are never listed there: neither needs a property to draw its items, so a gallery whose cover property is gone reports the missing cover and still shows every item.",
        "operationId": "GetContainerViews",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerViewsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Structure"
        ],
        "summary": "Replace the views a container offers",
        "description": "A whole-set replacement, because the order is part of what is being edited. A view's kind is one of 'list', 'board', 'calendar' or 'gallery'. What a kind must name is checked here - a board needs a grouping property, a calendar a date property - but whether that property exists is not: a view may be configured before the property is declared, and the read path reports the mismatch instead. Fails with 'views.invalid' when a view is not storable.",
        "operationId": "SetContainerViews",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetViewsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerViewsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContainerViewsResponse": {
        "required": [
          "views",
          "unrenderable",
          "default"
        ],
        "type": "object",
        "properties": {
          "views": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ViewResponse"
            }
          },
          "unrenderable": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "default": {
            "type": "string"
          }
        }
      },
...
      "EffectiveSchemaResponse": {
        "required": [
          "properties",
          "declared",
          "inherit"
        ],
        "type": "object",
        "properties": {
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyDefinitionResponse"
            }
          },
          "declared": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyDefinitionResponse"
            }
          },
          "inherit": {
            "type": "boolean"
          }
        }
      },
...
      "ItemResponse": {
        "required": [
          "id",
          "workspaceId",
          "parentId",
          "type",
          "title",
          "hasChildren",
          "seq",
          "lifecycleState",
          "properties",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "hasChildren": {
            "type": "boolean"
          },
          "seq": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "lifecycleState": {
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JsonObject": {
        "type": "object",
        "additionalProperties": { }
      },
...
      "ProblemDetails": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, namespaced by feature (for example 'health.check_not_found'). Clients branch on this value and never on 'title' or 'detail', which are human-facing and may change."
          },
          "traceId": {
            "type": "string",
            "description": "Identifier of the server-side trace that produced this failure."
          }
        }
      },
      "PropertyDefinitionRequest": {
        "required": [
          "key",
          "label",
          "type",
          "options",
          "required"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "options": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "required": {
            "type": "boolean"
          }
        }
      },
      "PropertyDefinitionResponse": {
        "required": [
          "key",
          "label",
          "type",
          "options",
          "required"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "required": {
            "type": "boolean"
          }
        }
      },
...
      "SetPropertiesRequest": {
        "required": [
          "properties"
        ],
        "type": "object",
        "properties": {
          "properties": {
            "$ref": "#/components/schemas/JsonObject"
          }
        }
      },
      "SetSchemaRequest": {
        "required": [
          "properties",
          "inherit"
        ],
        "type": "object",
        "properties": {
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyDefinitionRequest"
            }
          },
          "inherit": {
            "type": "boolean"
          }
        }
      },
      "SetViewsRequest": {
        "required": [
          "views",
          "default"
        ],
        "type": "object",
        "properties": {
          "views": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ViewRequest"
            }
          },
          "default": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ViewRequest": {
        "required": [
          "id",
          "name",
          "kind",
          "columns",
          "groupBy",
          "groupOrder",
          "dateProperty",
          "sortBy",
          "sortDescending",
          "mode",
          "coverProperty"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "columns": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "groupBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "groupOrder": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "dateProperty": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortDescending": {
            "type": "boolean"
          },
          "mode": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverProperty": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ViewResponse": {
        "required": [
          "id",
          "name",
          "kind",
          "columns",
          "groupBy",
          "groupOrder",
          "dateProperty",
          "sortBy",
          "sortDescending",
          "mode",
          "coverProperty"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "groupBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "groupOrder": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dateProperty": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "sortDescending": {
            "type": "boolean"
          },
          "mode": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverProperty": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}